requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.6.0
    # via django
backports-zoneinfo==0.2.1
    # via django
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
dj-database-url==0.5.0
    # via -r requirements.in
django==4.1
    # via
    #   -r requirements.in
    #   djangorestframework
    #   drf-yasg
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg==1.20.0
    # via -r requirements.in
exceptiongroup==1.1.1
    # via pytest
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-yasg
iniconfig==2.0.0
    # via pytest
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
markupsafe==2.1.2
    # via jinja2
packaging==23.1
    # via
    #   drf-yasg
    #   pytest
pluggy==1.0.0
    # via pytest
psycopg2-binary==2.9.3
    # via -r requirements.in
pytest==7.3.1
    # via pytest-django
pytest-django==4.5.2
    # via -r requirements.in
pytz==2023.3
    # via
    #   -r requirements.in
    #   djangorestframework
requests==2.30.0
    # via coreapi
ruamel-yaml==0.17.22
    # via drf-yasg
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via pytest
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==2.0.2
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
